chore: reorganize TargetsTable
