+++ ui/app/src/components/Scope/TargetsTable/PodsTable.tsx
@@ -21,7 +21,7 @@ import PaperContainer from '@ui/mui-extends/esm/PaperContainer'
 
 import { T } from 'components/T'
 
-import { TargetsTableActions } from './TargetsTable'
+import { TargetsTableActions } from '.'
 
 interface PodsTableProps extends TargetsTableActions {
   data: TypesPod[]
