+++ ui/app/src/components/Scope/TargetsTable/index.tsx
